Buffer all output from extractLocStrings before decoding
We decode after rather than in the subprocess call because it seems that extractLocStrings can occasionally flush its buffer without writing the entirety of a character out. When that happens, the text wrapper tries to decode it and fails. By buffering all bytes until the end of the command and then decoding, we avoid this issue.